Replace integer division with floor division
In Python 2, integer division is always treated as floor division, i.e.
1 / 2 = 0. Python 3 has changed this behavior, and 1 / 2 = 0.5. There
are some places where our code will not be happy with this change, so
switch all integer divisions to floor divisions (//), that has the same
behaviour in Python 2 and Python 3, and is the same with Python 2's
plain division when it comes to integers.

Note: possible occurrences of plain integer division where found by
traversing the Python AST using the following script:

 import sys
 import ast

 class DivPrinter(ast.NodeVisitor):
   def __init__(self, name):
     self.name = name
     super(ast.NodeVisitor, self).__init__()

   def visit_BinOp(self, node):
     if isinstance(node.op, ast.Div):
       # Skip divisions of/by floats, as the behavior didn't change
       for lr in (node.left, node.right):
         if isinstance(lr, ast.Num) and isinstance(lr.n, float):
           break
         if isinstance(lr, ast.Call) and getattr(lr.func, 'id', None) == 'float':
           break
       else:
         print("%s:%d" % (self.name, node.lineno))
     self.generic_visit(node)

 def main():
   for name in sys.argv[1:]:
     visitor = DivPrinter(name)
     with open(name, "r") as f:
       input_ = f.read()

     tree = ast.parse(input_, name)
     visitor.visit(tree)

 if __name__ == "__main__":
   main()

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>