Import print() from __future__ where necessary
While Python 2.6+ provides print() as a function, it is not the same
function as that used in Python 3 and does not support keyword arguments
(e.g. file=). In modules that use print(…, file=sys.stderr) we need to
import print_function from __future__ for the code to continue working.

Note that this is a stopgap to allow Python code to work with Python 2,
until the conversion to Python 3 is finished.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>