Fix NiceSort
NiceSort is supposed to split a string into chunks of digits and
non-digits, and then sort by them, sorting chunks of digits as numbers
and strings lexically.
But since not every string has digits and non-digits at the same
position, this leads to comparisons of strings and ints, which are
allowed (although nonsense) in python2 and disallowed in python3.
Also, there might be less than eight chunks, resulting in None.
Comparing None to numbers is also allowed in python2 but disallowed in
python3.

This tries to resurrect NiceSort by mapping strings to numbers as
-sum(map(ord, string)), which should always end up before the (positive)
numbers, and filtering out falsy values.

But overall, this function is rather crazy and should be replaced by
something sane.