Do not test to bulk remove the master from the master candidates
I do not think this should work.
The test gets a list of three master candidates and tries to bulk remove them.
Now, one of them is the master node. Removing it from the master
candidates should not be allowed and does in fact not work.
I guess that since the nodes are a dictionary, in python2 their order
was different, and now, since dicts are ordered, the master (node2) will
always be one of the first three MC nodes.