Exactly mimic the python2 comparison behaviour in NiceSort
NiceSort splits strings into groups of digits and non-digits, casting
the groups of digits to ints. There might also be Nones if the strings
do not contain enough such groups.

Comparing different types in python2 compares their type names. To mimic
that, we return tuples (typename, value). If the typename is different,
we will never compare the values (which could not be compared).